# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "malapos",
  "version": "0.0.1"
 },
 "paths": {
...
  "/api/v1/categories": {
   "get": {
    "operationId": "getCategories",
    "summary": "List categories",
    "tags": [
     "categories"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/categories.ts:18",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "post": {
    "operationId": "postCategories",
    "summary": "Create a category",
    "tags": [
     "categories"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/categories.ts:37",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "body"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 80
         },
         "sortOrder": {
          "type": "integer",
          "minimum": 0,
          "default": 0
         },
         "isActive": {
          "type": "boolean"
         }
        },
        "required": [
         "name"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/categories/{id}": {
   "patch": {
    "operationId": "patchCategoriesId",
    "summary": "Update a category",
    "tags": [
     "categories"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/categories.ts:75",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "body.partial()"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 80
         },
         "sortOrder": {
          "type": "integer",
          "minimum": 0,
          "default": 0
         },
         "isActive": {
          "type": "boolean"
         }
        },
        "additionalProperties": false
       }
      }
     }
    }
   },
   "delete": {
    "operationId": "deleteCategoriesId",
    "summary": "Delete a category",
    "tags": [
     "categories"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/categories.ts:89",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/categories/reorder": {
   "post": {
    "operationId": "postCategoriesReorder",
    "summary": "Create a reorder",
    "tags": [
     "categories"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/categories.ts:56",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "reorderCategoriesSchema"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "ids": {
          "type": "array",
          "items": {
           "type": "string",
           "minLength": 1
          },
          "minItems": 1,
          "maxItems": 500
         }
        },
        "required": [
         "ids"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
...
 },
...
}